Move tests for LoopRenamer
